Split off EntranceElement header
