+++ src/openrct2-ui/windows/Ride.cpp
@@ -64,6 +64,7 @@
 #include <openrct2/sprites.h>
 #include <openrct2/windows/Intent.h>
 #include <openrct2/world/Park.h>
+#include <openrct2/world/tile_element/EntranceElement.h>
 #include <optional>
 #include <string>
 #include <string_view>
