+++ src/openrct2/EditorObjectSelectionSession.cpp
@@ -30,6 +30,7 @@
 #include "windows/Intent.h"
 #include "world/Footpath.h"
 #include "world/Scenery.h"
+#include "world/tile_element/EntranceElement.h"
 
 #include <iterator>
 #include <vector>
