+++ src/openrct2/actions/FootpathLayoutPlaceAction.cpp
@@ -23,6 +23,7 @@
 #include "../world/Park.h"
 #include "../world/Surface.h"
 #include "../world/Wall.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/Slope.h"
 
 using namespace OpenRCT2;
