+++ src/openrct2/actions/FootpathPlaceAction.cpp
@@ -29,6 +29,7 @@
 #include "../world/Surface.h"
 #include "../world/TileElementsView.h"
 #include "../world/Wall.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/Slope.h"
 
 #include <algorithm>
