+++ src/openrct2/entity/Guest.cpp
@@ -60,6 +60,7 @@
 #include "../world/Scenery.h"
 #include "../world/Surface.h"
 #include "../world/TileElementsView.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "Peep.h"
 #include "Staff.h"
