+++ src/openrct2/entity/Staff.cpp
@@ -43,6 +43,7 @@
 #include "../world/Footpath.h"
 #include "../world/Scenery.h"
 #include "../world/Surface.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/Slope.h"
 #include "PatrolArea.h"
 #include "Peep.h"
