+++ src/openrct2/libopenrct2.vcxproj
@@ -599,6 +599,7 @@
     <ClInclude Include="Version.h" />
     <ClInclude Include="windows\Intent.h" />
     <ClInclude Include="windows\TileInspectorGlobals.h" />
+    <ClInclude Include="world\tile_element\EntranceElement.h" />
     <ClInclude Include="world\tile_element\TileElementType.h" />
     <ClInclude Include="world\tile_element\Slope.h" />
     <ClInclude Include="world\Banner.h" />
