+++ src/openrct2/rct1/S4Importer.cpp
@@ -72,6 +72,7 @@
 #include "../world/Surface.h"
 #include "../world/TilePointerIndex.hpp"
 #include "../world/Wall.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "RCT1.h"
 #include "Tables.h"
