+++ src/openrct2/ride/Ride.cpp
@@ -62,6 +62,7 @@
 #include "../world/Park.h"
 #include "../world/Scenery.h"
 #include "../world/TileElementsView.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "CableLift.h"
 #include "RideAudio.h"
 #include "RideConstruction.h"
