+++ src/openrct2/ride/TrackDesign.cpp
@@ -58,6 +58,7 @@
 #include "../world/Scenery.h"
 #include "../world/Surface.h"
 #include "../world/Wall.h"
+#include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/Slope.h"
 #include "Ride.h"
 #include "RideData.h"
