+++ src/openrct2/scripting/bindings/world/ScTileElement.cpp
@@ -22,6 +22,7 @@
 #    include "../../../world/Footpath.h"
 #    include "../../../world/Scenery.h"
 #    include "../../../world/Surface.h"
+#    include "../../../world/tile_element/EntranceElement.h"
 #    include "../../Duktape.hpp"
 #    include "../../ScriptEngine.h"
