+++ src/openrct2/world/Entrance.cpp
@@ -30,6 +30,7 @@
 #include "Map.h"
 #include "MapAnimation.h"
 #include "Park.h"
+#include "tile_element/EntranceElement.h"
 
 using namespace OpenRCT2;
