+++ test/tests/TileElements.cpp
@@ -17,6 +17,7 @@
 #include <openrct2/ParkImporter.h>
 #include <openrct2/world/Footpath.h>
 #include <openrct2/world/Map.h>
+#include <openrct2/world/tile_element/EntranceElement.h>
 
 using namespace OpenRCT2;
